piutan final v01

+++ database/migrations/2021_02_01_021842_create_detail_piutang_table.php
@@ -18,7 +18,7 @@ class CreateDetailPiutangTable extends Migration
             $table->string('transaksi_id');
             $table->timestamp('tanggal');
             $table->string('kasir_id');
-            $table->float('uang');
+            $table->integer('uang');
             $table->timestamps();
             $table->softDeletes();
         });
